requirements.txt
flask
boto3
pymysql
python-dotenv